+++ src/main/java/vectorwing/farmersdelight/common/CommonSetup.java
@@ -58,12 +58,14 @@ public class CommonSetup
 		ComposterBlock.COMPOSTABLES.put(ModItems.STRAW.get(), 0.3F);
 		ComposterBlock.COMPOSTABLES.put(ModItems.CABBAGE_SEEDS.get(), 0.3F);
 		ComposterBlock.COMPOSTABLES.put(ModItems.TOMATO_SEEDS.get(), 0.3F);
-		ComposterBlock.COMPOSTABLES.put(ModItems.RICE.get(), 0.65F);
-		ComposterBlock.COMPOSTABLES.put(ModItems.RICE_PANICLE.get(), 0.65F);
+		ComposterBlock.COMPOSTABLES.put(ModItems.RICE.get(), 0.3F);
+		ComposterBlock.COMPOSTABLES.put(ModItems.RICE_PANICLE.get(), 0.3F);
+		ComposterBlock.COMPOSTABLES.put(ModItems.SANDY_SHRUB.get(), 0.3F);
 
 		// 50% chance
-		ComposterBlock.COMPOSTABLES.put(ModItems.PUMPKIN_SLICE.get(), 0.65F);
-		ComposterBlock.COMPOSTABLES.put(ModItems.CABBAGE_LEAF.get(), 0.65F);
+		ComposterBlock.COMPOSTABLES.put(ModItems.PUMPKIN_SLICE.get(), 0.5F);
+		ComposterBlock.COMPOSTABLES.put(ModItems.CABBAGE_LEAF.get(), 0.5F);
+		ComposterBlock.COMPOSTABLES.put(ModItems.KELP_ROLL_SLICE.get(), 0.5F);
 
 		// 65% chance
 		ComposterBlock.COMPOSTABLES.put(ModItems.CABBAGE.get(), 0.65F);
@@ -88,13 +90,14 @@ public class CommonSetup
 		ComposterBlock.COMPOSTABLES.put(ModItems.CHOCOLATE_PIE_SLICE.get(), 0.85F);
 		ComposterBlock.COMPOSTABLES.put(ModItems.RAW_PASTA.get(), 0.85F);
 		ComposterBlock.COMPOSTABLES.put(ModItems.ROTTEN_TOMATO.get(), 0.85F);
+		ComposterBlock.COMPOSTABLES.put(ModItems.KELP_ROLL.get(), 0.85F);
 
 		// 100% chance
 		ComposterBlock.COMPOSTABLES.put(ModItems.APPLE_PIE.get(), 1.0F);
 		ComposterBlock.COMPOSTABLES.put(ModItems.SWEET_BERRY_CHEESECAKE.get(), 1.0F);
 		ComposterBlock.COMPOSTABLES.put(ModItems.CHOCOLATE_PIE.get(), 1.0F);
 		ComposterBlock.COMPOSTABLES.put(ModItems.DUMPLINGS.get(), 1.0F);
-		ComposterBlock.COMPOSTABLES.put(ModItems.STUFFED_PUMPKIN.get(), 1.0F);
+		ComposterBlock.COMPOSTABLES.put(ModItems.STUFFED_PUMPKIN_BLOCK.get(), 1.0F);
 		ComposterBlock.COMPOSTABLES.put(ModItems.BROWN_MUSHROOM_COLONY.get(), 1.0F);
 		ComposterBlock.COMPOSTABLES.put(ModItems.RED_MUSHROOM_COLONY.get(), 1.0F);
 	}